refactor : signUp refactor use viewmodel
